+++ test.js
@@ -5,6 +5,7 @@
  */
 
 const assert = require('assert')
+const g_secret = require('./index')
 
 /**
  * Assertions
@@ -13,5 +14,8 @@ const assert = require('assert')
 describe("index.js", () => {
   it("should have tests", () => {
     assert.ok(true)
+  })
+  it("should use SHA512, which uses 64-bit words", () => {
+    assert.equal((g_secret.length / 2), 64)
   })
 })